+++ CHANGES.md
@@ -5,6 +5,7 @@
 - speedup of model creation
 - add option to use persistent solver
 - change from pyomo AbstractModel to ConcreteModel formulation
+- introduce more tests
 
 ## v1.1.4
 2023-03-06